use Selenide instead Selenium in automation tests
