+++ build.gradle
@@ -20,8 +20,6 @@ idea {
 
 dependencies {
     compile 'junit:junit:4.12'
-    compile 'org.seleniumhq.selenium:selenium-java:2.48.2'
-    compile 'org.hamcrest:hamcrest-all:1.3'
     compile 'com.codeborne:selenide:3.0'
 }
