+++ src/test/java/com/intelliarts/conflab/automation/web/AddSpeakerView.java
@@ -4,6 +4,9 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
+import static com.codeborne.selenide.Condition.type;
+import static com.codeborne.selenide.Condition.visible;
+import static com.codeborne.selenide.Selenide.$;
 import static com.intelliarts.conflab.utils.ElementLocatorData.ADD_SPEAKER_BUTTON;
 import static com.intelliarts.conflab.utils.ElementLocatorData.CANCEL_BUTTON;
 import static com.intelliarts.conflab.utils.ElementLocatorData.OK_BUTTON;
@@ -12,8 +15,6 @@ import static com.intelliarts.conflab.utils.ElementLocatorData.SPEAKER_ABOUT;
 import static com.intelliarts.conflab.utils.ElementLocatorData.SPEAKER_EMAIL;
 import static com.intelliarts.conflab.utils.ElementLocatorData.SPEAKER_NAME;
 import static com.intelliarts.conflab.utils.ElementLocatorData.SPEAKER_POSITION;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.core.Is.is;
 
 public class AddSpeakerView extends BasicTestCase {
 
@@ -21,8 +22,8 @@ public class AddSpeakerView extends BasicTestCase {
     public static void setUp() throws Exception {
         basicSetUp();
         login();
-        driver.findElement(SPEAKERS_LINK).click();
-        driver.findElement(ADD_SPEAKER_BUTTON).click();
+        $(SPEAKERS_LINK).click();
+        $(ADD_SPEAKER_BUTTON).click();
     }
 
     @AfterClass
@@ -32,41 +33,41 @@ public class AddSpeakerView extends BasicTestCase {
 
     @Test
     public void nameFieldIsDisplayed() throws Exception {
-        assertThat(driver.findElement(SPEAKER_NAME).isDisplayed(), is(true));
+        $(SPEAKER_NAME).shouldBe(visible);
     }
 
     @Test
     public void emailFieldIsDisplayed() throws Exception {
-        assertThat(driver.findElement(SPEAKER_EMAIL).isDisplayed(), is(true));
+        $(SPEAKER_EMAIL).shouldBe(visible);
     }
 
     @Test
     public void emailFieldHasEmailType() throws Exception {
-        assertThat(driver.findElement(SPEAKER_EMAIL).getAttribute("type"), is("email"));
+        $(SPEAKER_EMAIL).shouldHave(type("email"));
     }
 
     @Test
     public void positionFieldIsDisplayed() throws Exception {
-        assertThat(driver.findElement(SPEAKER_POSITION).isDisplayed(), is(true));
+        $(SPEAKER_POSITION).shouldBe(visible);
     }
 
     @Test
     public void aboutTextareaIsDisplayed() throws Exception {
-        assertThat(driver.findElement(SPEAKER_ABOUT).isDisplayed(), is(true));
+        $(SPEAKER_ABOUT).shouldBe(visible);
     }
 
     @Test
     public void okButtonIsDisplayed() throws Exception {
-        assertThat(driver.findElement(OK_BUTTON).isDisplayed(), is(true));
+        $(OK_BUTTON).shouldBe(visible);
     }
 
     @Test
     public void okButtonHasTypeSubmit() throws Exception {
-        assertThat(driver.findElement(OK_BUTTON).getAttribute("type"), is("submit"));
+        $(OK_BUTTON).shouldHave(type("submit"));
     }
 
     @Test
     public void cancelButtonIsDisplayed() throws Exception {
-        assertThat(driver.findElement(CANCEL_BUTTON).isDisplayed(), is(true));
+        $(CANCEL_BUTTON).shouldBe(visible);
     }
 }
