+++ src/test/java/com/intelliarts/conflab/automation/web/SpeakersView.java
@@ -1,14 +1,14 @@
 package com.intelliarts.conflab.automation.web;
 
+import com.codeborne.selenide.Condition;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
+import static com.codeborne.selenide.Selenide.$;
 import static com.intelliarts.conflab.utils.ElementLocatorData.ADD_SPEAKER_BUTTON;
 import static com.intelliarts.conflab.utils.ElementLocatorData.SPEAKERS_LINK;
 import static com.intelliarts.conflab.utils.ElementLocatorData.SPEAKERS_PAGE_HEADER;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.core.Is.is;
 
 public class SpeakersView extends BasicTestCase {
 
@@ -16,7 +16,7 @@ public class SpeakersView extends BasicTestCase {
     public static void setUp() throws Exception {
         basicSetUp();
         login();
-        driver.findElement(SPEAKERS_LINK).click();
+        $(SPEAKERS_LINK).click();
     }
 
     @AfterClass
@@ -26,11 +26,11 @@ public class SpeakersView extends BasicTestCase {
 
     @Test
     public void speakersHeader() throws Exception {
-        assertThat(driver.findElement(SPEAKERS_PAGE_HEADER).getText(), is("Speakers"));
+        $(SPEAKERS_PAGE_HEADER).shouldHave(Condition.text("Speakers"));
     }
 
     @Test
     public void addButtonVisible() throws Exception {
-        assertThat(driver.findElement(ADD_SPEAKER_BUTTON).isDisplayed(), is(true));
+        $(ADD_SPEAKER_BUTTON).shouldBe(Condition.visible);
     }
 }
